requirements.txt
streamlit
pandas
plotly
fastapi
uvicorn
python-jose[cryptography]
